refactor(meta): relocate meta package to pkg/openlore/meta
Home the frontmatter-scanning package under the openlore namespace. It stays a
separate, cmds-free package (deps: vfs, okf only), so pkg/shell/cmds can still
call it without an import cycle while the openlore host reuses it.
